pyproject.toml
[project]
name = "blog-api"
version = "0.1.0"
description = "Django REST Framework Blog API"
requires-python = ">=3.8"
dependencies = [
    "cryptography>=46.0.3",
    "django>=4.2.0",
    "django-allauth>=65.13.0",
    "django-cors-headers>=4.4.0",
    "django-filter>=24.3",
    "django-imagekit>=6.0.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt[crypto]>=5.3.1",
    "dotenv>=0.9.9",
    "pillow>=10.4.0",
    "python-decouple>=3.8",
    "requests>=2.32.4",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
]
